+++ Game/GamePlayerCpu.cpp
@@ -25,9 +25,7 @@ void GamePlayerCpu::PlayMyTurn(BOARD_INFO enBoardInfo)
 	std::vector<unsigned char> v(enBoardInfo.enSize.ucRow * enBoardInfo.enSize.ucCol);
 	std::iota(v.begin(), v.end(), 0);
 
-	unsigned seed = std::chrono::system_clock::now()
-		.time_since_epoch()
-		.count();
+	unsigned int seed = static_cast<unsigned int>(std::chrono::system_clock::now().time_since_epoch().count());
 	std::shuffle(std::begin(v), std::end(v), std::default_random_engine(seed));
 
 	DISC_MOVE enDiscMove = { GamePlayer::GetDiscCol(),{0,0} };
